Finish work on the set name feature
To rename a set, you can long-press on the name of the set in the
builder.

In order to deal with the new variable length set names, I broke the set
name and workout duration into two separate lines in the workout runner.
I don't love having to eat up more vertical space that way, but don't
have another amazing idea for how to deal with it and I didn't want to
shrink down the font size