Add support for spinner granularity and automatic text edit display
Spinner granularity can be set to 5, 10, or 15 seconds in the
preferences now.

If a time value is entered for an exercise that cannot be shown in the
spinner, then it will automatically show the text edit instead of the
spinner. Changing the value to a valid spinner value will call the
spinner to reappear.