Add tap to edit functionality to number spinners
This is optional functionality that you have to enable the pref to get.

Once it is set, you have to tap on a spinner to get it to activate. The
spinner will automatically deactivate when the user stops scrolling.
When the pref is turned off, the behavior should be exactly as it is
now.